feat: MHG-08 - MHG-11

+++ smartcontracts/src/hm-token/HMToken.sol
@@ -20,13 +20,13 @@ contract HMToken is HMTokenInterface {
 
     uint256 private constant MAX_UINT256 = ~uint256(0);
     uint256 private constant BULK_MAX_VALUE = 1000000000 * (10 ** 18);
-    uint32  private constant BULK_MAX_COUNT = 100;
+    uint32 private constant BULK_MAX_COUNT = 100;
 
     event BulkTransfer(uint256 indexed _txId, uint256 _bulkCount);
     event BulkApproval(uint256 indexed _txId, uint256 _bulkCount);
 
-    mapping (address => uint256) private balances;
-    mapping (address => mapping (address => uint256)) private allowed;
+    mapping(address => uint256) private balances;
+    mapping(address => mapping(address => uint256)) private allowed;
 
     string public name;
     uint8 public decimals;
@@ -54,7 +54,8 @@ contract HMToken is HMTokenInterface {
         balances[_spender] = balances[_spender].sub(_value, "Spender balance too low");
         balances[_to] = balances[_to].add(_value);
 
-        if (_allowance != MAX_UINT256) { // Special case to approve unlimited transfers
+        if (_allowance != MAX_UINT256) {
+            // Special case to approve unlimited transfers
             allowed[_spender][msg.sender] = allowed[_spender][msg.sender].sub(_value);
         }
 
@@ -74,11 +75,12 @@ contract HMToken is HMTokenInterface {
         return true;
     }
 
-    function increaseApproval(address _spender, uint _delta) public returns (bool success) {
+    function increaseApproval(address _spender, uint256 _delta) public returns (bool success) {
         require(_spender != address(0), "Token spender is an uninitialized address");
 
-        uint _oldValue = allowed[msg.sender][_spender];
-        if (_oldValue.add(_delta) < _oldValue || _oldValue.add(_delta) >= MAX_UINT256) { // Truncate upon overflow.
+        uint256 _oldValue = allowed[msg.sender][_spender];
+        if (_oldValue.add(_delta) < _oldValue || _oldValue.add(_delta) >= MAX_UINT256) {
+            // Truncate upon overflow.
             allowed[msg.sender][_spender] = MAX_UINT256.sub(1);
         } else {
             allowed[msg.sender][_spender] = allowed[msg.sender][_spender].add(_delta);
@@ -87,11 +89,12 @@ contract HMToken is HMTokenInterface {
         return true;
     }
 
-    function decreaseApproval(address _spender, uint _delta) public returns (bool success) {
+    function decreaseApproval(address _spender, uint256 _delta) public returns (bool success) {
         require(_spender != address(0), "Token spender is an uninitialized address");
 
-        uint _oldValue = allowed[msg.sender][_spender];
-        if (_delta > _oldValue) { // Truncate upon overflow.
+        uint256 _oldValue = allowed[msg.sender][_spender];
+        if (_delta > _oldValue) {
+            // Truncate upon overflow.
             allowed[msg.sender][_spender] = 0;
         } else {
             allowed[msg.sender][_spender] = allowed[msg.sender][_spender].sub(_delta);
@@ -104,18 +107,22 @@ contract HMToken is HMTokenInterface {
         return allowed[_owner][_spender];
     }
 
-    function transferBulk(address[] memory _tos, uint256[] memory _values, uint256 _txId) public override returns (uint256 _bulkCount) {
+    function transferBulk(address[] memory _tos, uint256[] memory _values, uint256 _txId)
+        public
+        override
+        returns (uint256 _bulkCount)
+    {
         require(_tos.length == _values.length, "Amount of recipients and values don't match");
         require(_tos.length < BULK_MAX_COUNT, "Too many recipients");
 
         uint256 _bulkValue = 0;
-        for (uint j = 0; j < _tos.length; ++j) {
+        for (uint256 j = 0; j < _tos.length; ++j) {
             _bulkValue = _bulkValue.add(_values[j]);
         }
         require(_bulkValue < BULK_MAX_VALUE, "Bulk value too high");
 
         bool _success;
-        for (uint i = 0; i < _tos.length; ++i) {
+        for (uint256 i = 0; i < _tos.length; ++i) {
             _success = transferQuiet(_tos[i], _values[i]);
             if (_success) {
                 _bulkCount = _bulkCount.add(1);
@@ -125,18 +132,21 @@ contract HMToken is HMTokenInterface {
         return _bulkCount;
     }
 
-    function approveBulk(address[] memory _spenders, uint256[] memory _values, uint256 _txId) public returns (uint256 _bulkCount) {
+    function approveBulk(address[] memory _spenders, uint256[] memory _values, uint256 _txId)
+        public
+        returns (uint256 _bulkCount)
+    {
         require(_spenders.length == _values.length, "Amount of spenders and values don't match");
         require(_spenders.length < BULK_MAX_COUNT, "Too many spenders");
 
         uint256 _bulkValue = 0;
-        for (uint j = 0; j < _spenders.length; ++j) {
+        for (uint256 j = 0; j < _spenders.length; ++j) {
             _bulkValue = _bulkValue.add(_values[j]);
         }
         require(_bulkValue < BULK_MAX_VALUE, "Bulk value too high");
 
         bool _success;
-        for (uint i = 0; i < _spenders.length; ++i) {
+        for (uint256 i = 0; i < _spenders.length; ++i) {
             _success = increaseApproval(_spenders[i], _values[i]);
             if (_success) {
                 _bulkCount = _bulkCount.add(1);
